timestampMicroService/requirements.txt
Django==3.2.9
django-cors-headers==3.13.0
djangorestframework==3.14.0
psycopg2==2.9.5
sqlparse==0.4.2